refactor: split EventMarkets into focused sub-components
Extracts 4 files from the 1267 LOC monolith (now 796 LOC):
- ResolvedMarketRow.tsx (resolved market row display)
- OtherOutcomeRow.tsx (other outcome row)
- MarketDetailTabs.tsx (market detail tabs section)
- eventMarketUtils.ts (shared helpers: isMarketResolved, resolveWinningOutcomeIndex)
